refactor: clean up -> reorder functions, cluster them by logic proximity
rename vars, function and file for better understading and matching
conventions.
